[inductor] Reinplacing should not allow an op to mutate the same input multiple times
Fixes #132196

Let's say we have:
- op(x, y) that mutates both x and y
- new_x, new_y = functional_op(x, y) is the functional variant
If we are presented with functional_op(x, x), we must not reinplace
this into op(x, x), because then it would be writing to the same Tensor.
Instead, it's OK to reinplace one of them and to clone the other:
>>> y = x.clone()
>>> op(x, y)
This also applies if we have views: functional_op(x, x[0])
should not reinplace into op(x, x[0]).

The fix is to avoid reinplacing an arg if a view of it already has been
reinplaced.

Test Plan:
- new and existing tests

ghstack-source-id: e7743718db4a1cc5f29ce2adeb6f4b8951f7c07e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/132238