[inductor] Fix test_codecache::test_inductor_counters
Summary: This test is flakey internally, but it's not a great test in the first place since it's relying on the max-autotune step to bump a related counter. Instead of doing that, directly install a mock that bumps a counter specifically for this test. Additionally, test that the caching logic correctly accommodates an arbitrary counter delta (previously the relevant counter is only bumped by +1).

ghstack-source-id: 85afeac659f5265cc48b304d8c84f967da5cd9ad
Pull Request resolved: https://github.com/pytorch/pytorch/pull/133244