[inductor] Skip the sub-process pool until it's ready
Summary: Torch-compiling a quick script can be a bit slower than it needs to be: even though we initialize the subprocess pool early, it still might not be ready by the time we try to compile the first Triton kernel. Instead, let's use the single-threaded path until the pool has successfully completed a no-op job.

ghstack-source-id: 205c0c2c9506421565c2b72ed7e5ec7446d750fb
Pull Request resolved: https://github.com/pytorch/pytorch/pull/133508