[inductor] Fix needs_fixed_stride_order silent incorrectness
Fixes #128084

The approach is two-fold (based on what Elias suggested in the comment
thread):
1. We realize Tensors and ensure they are the right stride.
2. If we can't ensure they have the right stride (if the tensors are
   already in an incorrect stride with FixedLayout, or if there's a
   view), then we require the correct stride on use via a clone and then
   copy_ back the result of the mutation.

We can improve this based on use cases (if the tensor is a view, then we
can require the base of the view to have the right stride order). The
view case isn't very common so I'm ignoring it for now.

Test Plan:
- tests

ghstack-source-id: ce50fe6657cb6ee345ded0fd54a2f7887915804b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/133638