[DCP] Fixes the BC issue where the traversal doesn't support versions before 2.4
The original DCP doesn't flattening all the containers, which can cause issues, https://github.com/pytorch/pytorch/pull/125335 intends to solve the issue by flattening all the dictionaries.

Unfortunately, it breaks the checkpoints that are saved before 2.4. This
also shows some issues of the DCP:

1. DCP should record version in the metadata.
2. DCP should have a nice way to load old state_dict.
3. DCP should unflatten all containers (map, list) not just map.

This PR only addresses issue 2 to unblock users. Issue 1 and issue 3 need to be addressed in the future.

ghstack-source-id: f207aed952b19138893cadfbe218363098e63e40
Pull Request resolved: https://github.com/pytorch/pytorch/pull/134158