[fp8 rowwise] Support transposing operands in order to change output layout
On some occasion, a column-major output layout is more efficient (it's unclear if it's because of better store coalescing for some tile shapes, or whether it's just that it's CUTLASS's default and thus it's better optimized).

At this stage I only add a flag that allows to transpose, but the hardest will be deciding on a new heuristic to turn it on selectively. This will be in a follow-up PR.

ghstack-source-id: 693566413386183b69361bb19f00348f173f6c8a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/134773