Fix tensor.data access under inference_mode and compile
Fixes https://github.com/pytorch/pytorch/issues/134798

In the regular Tensor case, when you call Tensor.data, there's a check
for if inference mode is active. If it is active, then we don't set the
version counter. We replicate this check for Tensor Subclasses (the bug
was we were trying to set the version counter on a FakeTensor in
inference_mode).

Test Plan:
- new test

ghstack-source-id: de6a178fbaf1631e381e006994ea6dc66d6222ec
Pull Request resolved: https://github.com/pytorch/pytorch/pull/134878