Add validator to ensure runner determinator script is kept in sync We keep two copies of the runner-determinator script: 1. In runner_determinator.py, for ease of testing.  This however is not actually executed during CI 2. Embedded in _runner-determinator.yml.  This is what CI uses.
Why the duplication? Short version: Because of how github CI works, during a given CI run the workflow yml files could actually come from the main branch, while the remaining files get read from the local commit.
This can lead to a newer version of _runner-determinator.yml trying to invoke an older version of runner_determintor.py than it was actually designed for. Chaos ensues.

We mitigate this by embedding the script into the yml file.  But we still keep the script around because it's much easier to run tests against.

This workflow's job is to ensure that if one edits the script in one of those two locations then they remember to update it in the other location as well

ghstack-source-id: fbe0a6e8a1efc1d4ce561c8283ad707886419712
Pull Request resolved: https://github.com/pytorch/pytorch/pull/134800