Make Dynamo inline through torch._library.custom_ops.autograd
Fixes https://github.com/pytorch/pytorch/issues/135057

The bug was: in the situation that Dynamo graph breaks in the forward
and Compiled Autograd uses Dynamo to introspect the backward, we end up
running into a "Unsupported: inlining through SKIPFILES" error. The
solution is to mark the entirety of this module as inlineable.

Test Plan:
- new test

ghstack-source-id: 63bfc873c9099f6fb028d65829b1b8aead69ce88
Pull Request resolved: https://github.com/pytorch/pytorch/pull/135066