[inductor][triton] mark workspace args as mutated
SplitScan makes use of a workspace arg that needs to be zeroed before it is used - then, it is used to communicate between thread blocks during the triton kernel implementation. It is mutated during during the execution of the kernel, so it should be marked as such.

Before this PR, it is not marked as mutated; AFAIK this is fine during normal execution, but during autotuning it causes problems. The workspace starts off zeroed (as expected), but during autotuning the kernel will be executed multiple times and the workspace does not get re-set between executions, resulting in device-side assert failures. The test added in this PR repros the issue when the fix is removed.

When we mark the arg as mutated, then the arg gets cloned before autotuning, so that the arg passed to the kernel during autotuning will always be zeroed as expected.
https://github.com/pytorch/pytorch/blob/804852c1f998e6c7c58a5c7cbcb5e9063d02d54a/torch/_inductor/runtime/triton_heuristics.py#L685-L689

ghstack-source-id: 11a58f71ab94f06a3a7d3d97e748cb6ee0a41159
Pull Request resolved: https://github.com/pytorch/pytorch/pull/134648